# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human readable message
        type: string
    type: object
...
  internal_features_users_transport_http_dto.CreateUserRequest:
    properties:
      full_name:
        example: Коля Колевич
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+77758451038"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http_dto.CreateUserResponse:
    properties:
      full_name:
        example: Колясик Колясикович
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79132038908"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http_dto.GetUserResponse:
    properties:
      full_name:
        example: Колясик Колясикович
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79132038908"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http_dto.PatchUserRequest:
    properties:
      full_name:
        example: Максим Максимыч
        type: string
      phone_number:
        example: "+77758451038"
        type: string
    type: object
  internal_features_users_transport_http_dto.PatchUserResponse:
    properties:
      full_name:
        example: Колясик Колясикович
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79132038908"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http_dto.UserDTOResponse:
    properties:
      full_name:
        example: Колясик Колясикович
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79132038908"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Просмотр списка пользователей с опциональной пагинацией
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http_dto.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http_dto.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаление существующего в системе пользователя по его ID
      parameters:
      - description: ID удаляемого пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Получение пользователя по его ID
      parameters:
      - description: ID получаемого пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно найден
          schema:
            $ref: '#/definitions/internal_features_users_transport_http_dto.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующем в системе пользователе
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `phone_number` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"phone_number": "+711122233344"` — устанавливает новый номер
        3. **Передан null**: `"phone_number": null` - очищает поле в БД (set to NULL)
        Ограничение: `fullname` не может быть выставлен как null
      parameters:
      - description: ID изменяемого пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http_dto.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённый пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http_dto.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Изменение пользователя
      tags:
      - users
swagger: "2.0"